Polymorphism can be inferred from arguments in some scenarios There are some cases that are incomplete. Arguments of struct type that are polymorphic are not inferred. Arguments that aren't polymorphic or are solely a polymorphic type will be inferred. This is because inferrence of poly. structs require a recursive method which is a little harder to implement. This will be tackled another day. I have left asserts in the mean time.
Methods can now refer to members without using the 'this' argument. Identifiers for each member of the struct type are implicitly made.

DirectoryIterator is now available when compiling x64 programs.

The stack wasn't properly aligned to 16 bytes in some cases which caused crashes when running programs from the x64 backend. This has now been fixed. Unfortunately, the interpreter was broken in the process.