Finally created a valid but empty PDB The function PDBFile::WriteEmpty is the only function that successfully writes PDB files at the moment. The purpose of the function is to act as a tempalte to look back on when things aren't working.
The class/struct PDBFile will be the interface for reading and writing PDB files. PDB files are complex so an interface will help with hiding that complexity.

Added DebugInformation which links instructions to source lines, functions and local variables. ObjectWriter will convert the information to a PDB file when PDBFile is completed.