You can now debug executables from the compiler with a C/C++ debugger
There are still some issues when stepping through the program line by line but the fact that you can use a debugger is fantastic. Being able to see the disassembly of the source code will really help with finding bugs.

- Fixed some failing test cases.
- Fixed memory leaks.

The old error messages have been replaced with the new ones. There are some unused code in comments which use ERR_HEAD2/3 but there is no point in replacing those if they will be deleted some time later.