Fixed a crash when running perf.btb
FnOverloads::getOverload was picking the wrong overload because an integer was allowed to be casted to a pointer. The overload with the pointer was choosen before the overload with the integer.

The current fix is to ensure that a function call only suits one overload when casting is allowed. Unless the types are integers.