Tests on Unix, better messages in test suite, bug with operator overloading, bug with defer, #quote in preprocessor,
Improved/Added:
- Printing of test suite is now more stylish.
- Started working on a reporter which should store error messages for later.
- #quote can be placed before a macro or token which will turn all tokens (from macro) or the single token into a double quoted token. (not fully tested, it seems to work)

Bug fixes:
- Overloading an operator forced every expression to try to use the overload instead of the normal behaviour. Normal behaviour is used if types don't match (or can't be converted).
- Defer in loops were not executed in the reverse order BUT now they are.

I found another bug in recur.btb with many expressions and operator overloading. The code type checks operator overloading and it's expressions recursively and if it fails it checks the expressions again resulting in 2**n type algorithm. 22 + operations in a statement takes can take 11 seconds which is crazy.