Added type information (Lang.btb)
Lang.btb contains global variables which are slices that points to the type information in the .data section of the object file.

Note that this is the first implementation and prototype of it. I have not tested whether it works with polymorphism and other special scenarios.

typeinfo.btb has a couple of tests but we need some more. Don't be lazy with tests!