Fixed bug when disabling code generation (TEST_ERROR, no_code)
The annotations TEST_ERROR, no_code can be used to disable code generation for a statement or scope. The state of the stack was not reset after the generator was running in the disabled code generation mode. Now we do reset the state of the stack to how it was before we encountered the no code statments/scopes.

The reason the bug wasn't found earlier was because the test suite didn't check the exit code causing it to miss the access violations.

Coincidentally, we ran all tests before the access violation was encountered so we didn't notice anything there either.