Fixed type information, globals variables, and more bugs
Changed TASK_TYPE_BODIES to TASK_TYPE_BODY.
Instead of one task for all function bodies in an import we create many
a task per function body to check. Previously we had a
global list for all func impl to check which meant that any
import could check function for any other function. This is bad
if there are global variables that haven't been checked yet.