Fixed inline assembly and callee saved registers
@stdcall and @unixcall functions will save non-volatile registers (rbx, rsi, rdi, r12, r13...).
The language's own calling convention does not save those non-volatile registers.

When x64_gen is better optimized to know which non-volatile registers are used then we can always save them but x64_gen is not that smart and will just save all those registers. It also becomes hard to look at disassembly with the extra push and pops so it's nicer to not save them unless you really need it.