Fixes to Linux version, pthreads bindings, better function overload matching
Finding the right overload for function calls (CheckFncall in TypeChecker.cpp) was not done correctly. A function name that matched with a method would only look at the method's overloads and ignore normal functions. This is fixed.

Besides pthreads bindings, target independent wrapper for sockets was implemented (Sockets.btb) as well as fixes to Net.btb so that it works on Linux too.

Fixed bug in CheckType where function pointers aren't handled. Argument types that are polymorphic were not created, a pointer as return value was seen as a pointer to the function pointer itself.