Added GLFW joystick bindings, collision for the crawler game
Fixed bug where enums in methods were skipped

Fixed a crash with x64 generator, something with alloc_local

Fixed unary subtraction for floats, xor on the sign bit was emitted in for the bytecode but the x64 generator crashed because it didn't handle it. Now we do subtraction with a zero instead of messing with bits.

Added automatic detection for controllers (glfw joysticks) where a player is spawned for each connected controller (player is removed if controller disconnects).

Added automatic camera movement and zooming so that all local players are visible at all times.

Added pause and frame step features from the other blob game example.

Added saving and loading of worlds to and from files.

Redid the collision code for entities. It's still not great, gaps between entities after collision, but it works better than before.