Added chat to crawler game, altough it crashes when starting server and debug info isn't work
I would normally fix the bug before commiting but the since debug info isn't working and i haven't tested it for a while, it could take a long time to fix. I commit so I can go back to earlier commits and see where it broke, that way i don't have to stash the changes.