Little work on exceptions, changes to code generation
This work on exceptions only apply for Windows. We will have to do a similar thing on Linux (using DWARF or sigaction).

With exceptions you need to unwind the function calls and the stack when searching for the appropriate catch block. To unwind the stack you need specify how in the .xdata section. This is done by unwind codes which describes the prolog of the function. Some codes save non-volatiles registers, others allocates local variables (by incrementing RSP).

Currently we can easily specify that RBP is pushed and saved on the stack, RSP however is more complicated due to the code generation. The generator increments and decrements RSP all over the place in the function and even inside conditional jumps. The unwind codes cannot represent this so we must limit the increment and decrement of rsp to the prolog and epilog of the function. That is why this commit modifies the generator to deal with this.

I am not done with the change yet. Most tests fail and there are bugs to be fixed.