Try catch works with exceptions in the direct scope
Try-catch does not work if the exceptions occurs in a called function. The exception handler does not unwind the stack to find a try-catch block, it just goes through the blocks of the current function.

Another missing feature is filtering exceptions. The expression after 'catch' is ignored. All exceptions are always caught which is not what we want.