Exception handler unwinds stack and filter codes
Exceptions are pretty much done and usable on Windows. What remains is a couple of extra convenient features and test cases. Perhaps a little documentation for the future for anyone looking to improve or reimplement the exception system.

Added:
- Exception handler will now look at handler data provided in .xdata section for the appropriate function iteratively while traversing up the stack.
- The exception handler will properly restore the stack. Before, the stack would have been messed up because of remaining pushes and argument allocations before the exception occured.
- The catch filter expressions are no longer ignored. You can specify to catch only division by zero errors if you want or specify all the exceptions.

Fixed some x64 and generator bugs with casting signedness, parameters, and break/continue with new stack allocations/management.