Fixed bug with in x64 gen with function stack alignment
The x64 generator had a bug with keeping track of the virtual stack pointer.
We use stack pointer moments to save and restore the virtual pointer when emitting jump instructions. That way, code before and after if and while blocks will have the correct virtual stack pointer.

The bug was that we didn't save stack moments for normal jump instructions, only conditional jumps. Basically, return statements in 'else' scopes will ruin the calculations for 16-byte alignment which is required before function calls.