Started on simple platform independent Audio API
This only works on Linux in this commit. I will fix Windows in a later commit.

This Audio API is seperate from the SoundSystem i was working on earlier.

SoundSystem is for advanced mixing while Audio AP is about read and writing audio data from devices while  Linux and Windows specifics are hidden.